# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCP Tool Evaluation System API
  description: |
    REST API for evaluating Model Context Protocol (MCP) tool effectiveness.
    
    **Features**:
    - Tool selection evaluation with LLM providers
    - Parameter validation (completeness, correctness, type conformance)
    - Inter-tool similarity detection and recommendations
    - Comprehensive metrics (precision, recall, F1, parameter accuracy)
    - Security-first architecture (runtime API keys, never persisted)
    - Fresh tool ingestion per test run for accuracy
    
    **Generated**: 2025-11-10
    **Git Commit**: 859d728254fc64586c05039908056c6b4bec1709
  version: 2.0.0
  contact:
    name: mcp-tef Team
    url: https://github.com/StacklokLabs/mcp-tef
servers:
  - url: https://localhost:8000
    description: Local development server (HTTPS with auto-generated cert)
  - url: http://localhost:8000
    description: Local development server (HTTP, use --no-tls-enabled)
...
paths:
...
  /similarity/analyze:
    post:
      tags: [similarity]
      summary: Analyze tool similarity
      description: |
        **002-inter-tool-similarity**: Full similarity analysis with embedding-based detection.
        Supports server_list, tool_list, or url_list formats (stateless, no database storage).
      operationId: analyzeSimilarity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimilarityAnalysisRequest'
      responses:
        '200':
          description: Similarity analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarityAnalysisResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
  /similarity/matrix:
    post:
      tags: [similarity]
      summary: Generate similarity matrix
      description: Returns complete pairwise similarity matrix for all tools
      operationId: generateSimilarityMatrix
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimilarityAnalysisRequest'
      responses:
        '200':
          description: Similarity matrix generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarityMatrixResponse'
  /similarity/overlap-matrix:
    post:
      tags: [similarity]
      summary: Generate capability overlap matrix
      description: Multi-dimensional overlap analysis (semantic, parameters, description)
      operationId: generateOverlapMatrix
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimilarityAnalysisRequest'
      responses:
        '200':
          description: Overlap matrix generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverlapMatrixResponse'
  /similarity/recommendations:
    post:
      tags: [similarity]
      summary: Get differentiation recommendations
      description: |
        Analyze tool pair and generate actionable differentiation recommendations.
        Requires exactly 2 tools in request.
      operationId: getDifferentiationRecommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimilarityAnalysisRequest'
      responses:
        '200':
          description: Recommendations generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DifferentiationRecommendationResponse'
        '400':
          description: Must provide exactly 2 tools

# ==============================================================================
# COMPONENTS
# ==============================================================================
components:
...
  schemas:
...
    SimilarityAnalysisRequest:
      type: object
      properties:
        server_list:
          type: array
          items:
            type: object
          nullable: true
          description: 'Array of MCP servers with embedded tools'
        tool_list:
          type: array
          items:
            type: object
          nullable: true
          description: 'Array of tool definitions (direct)'
        url_list:
          type: array
          items:
            type: string
          nullable: true
          description: 'Array of MCP server URLs to fetch from'
        embedding_model:
          type: string
          nullable: true
        similarity_threshold:
          type: number
          format: float
          default: 0.85
          minimum: 0.0
          maximum: 1.0
        compute_full_similarity:
          type: boolean
          default: false
        include_recommendations:
          type: boolean
          default: false
    SimilarityAnalysisResponse:
      type: object
      required: [tool_ids, matrix, threshold, flagged_pairs, generated_at]
      properties:
        tool_ids:
          type: array
          items:
            type: string
        matrix:
          type: array
          items:
            type: array
            items:
              type: number
              format: float
        threshold:
          type: number
          format: float
        flagged_pairs:
          type: array
          items:
            type: object
        generated_at:
          type: string
          format: date-time
        recommendations:
          type: array
          items:
            type: object
          nullable: true
    SimilarityMatrixResponse:
      type: object
      required: [tool_ids, matrix, threshold, generated_at]
      properties:
        tool_ids:
          type: array
          items:
            type: string
        matrix:
          type: array
          items:
            type: array
            items:
              type: number
              format: float
        threshold:
          type: number
          format: float
        flagged_pairs:
          type: array
          items:
            type: object
        generated_at:
          type: string
          format: date-time
    OverlapMatrixResponse:
      type: object
      required: [tool_ids, matrix, dimensions, generated_at]
      properties:
        tool_ids:
          type: array
          items:
            type: string
        matrix:
          type: array
          items:
            type: array
            items:
              type: number
              format: float
        dimensions:
          type: object
          properties:
            semantic:
              type: number
              format: float
            parameters:
              type: number
              format: float
            description:
              type: number
              format: float
        generated_at:
          type: string
          format: date-time
    DifferentiationRecommendationResponse:
      type: object
      required: [tool_pair, similarity_score, issues, recommendations, generated_at]
      properties:
        tool_pair:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 2
        similarity_score:
          type: number
          format: float
        issues:
          type: array
          items:
            type: object
        recommendations:
          type: array
          items:
            type: object
        generated_at:
          type: string
          format: date-time

    # ==========================================================================
    # ERROR RESPONSES
    # ==========================================================================
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security: []